Move ThreadedFeedAccess out as independent class.
